Split into packages and add EndpointConfig
