--- src/test/java/autotests/FlyTest.java
@@ -1,101 +1,0 @@
-package autotests;
-
-import com.consol.citrus.TestCaseRunner;
-import com.consol.citrus.annotations.CitrusResource;
-import com.consol.citrus.annotations.CitrusTest;
-import com.consol.citrus.testng.spring.TestNGCitrusSpringSupport;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.testng.annotations.Optional;
-import org.testng.annotations.Test;
-
-import static com.consol.citrus.dsl.MessageSupport.MessageBodySupport.fromBody;
-import static com.consol.citrus.http.actions.HttpActionBuilder.http;
-
-public class FlyTest extends TestNGCitrusSpringSupport {
-    @Test(description = "Проверка того, что уточка летает с существующим id и активными крыльями")
-    @CitrusTest
-    public void flyWithActiveWings(@Optional @CitrusResource TestCaseRunner runner) {
-        createDuck(runner, "yellow", 0.03, "rubber", "quack", "ACTIVE");
-        getDuckId(runner);
-        duckFly(runner, "${duckId}");
-        validateResponse(runner, "{\n"+ "  \"message\": \"I am flying\"\n"+ "}");
-        deleteDuck(runner, "${duckId}");
-    }
-
-    @Test(description = "Проверка того, что уточка не летает с существующим id и со связанными крыльями")
-    @CitrusTest
-    public void flyWithFixedWings(@Optional @CitrusResource TestCaseRunner runner) {
-        createDuck(runner, "yellow", 0.03, "rubber", "quack", "FIXED");
-        getDuckId(runner);
-        duckFly(runner, "${duckId}");
-        validateResponse(runner, "{\n"+ "  \"message\": \"I can not fly\"\n"+ "}");
-        deleteDuck(runner, "${duckId}");
-    }
-
-    @Test(description = "Проверка того, что уточка не летает с существующим id и крыльями в неопределенном состоянии")
-    @CitrusTest
-    public void flyWithUndefinedWings(@Optional @CitrusResource TestCaseRunner runner) {
-        createDuck(runner, "yellow", 0.03, "rubber", "quack", "UNDEFINED");
-        getDuckId(runner);
-        duckFly(runner, "${duckId}");
-        validateResponse(runner, "{\n"+ "  \"message\": \"Wings are not detected\"\n"+ "}");
-        deleteDuck(runner, "${duckId}");
-    }
-
-    public void duckFly(TestCaseRunner runner, String id) {
-        runner.$(
-                http()
-                        .client("http://localhost:2222")
-                        .send()
-                        .get("/api/duck/action/fly")
-                        .queryParam("id", id));
-    }
-    public void createDuck(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
-        runner.$(
-                http()
-                        .client("http://localhost:2222")
-                        .send()
-                        .post("/api/duck/create")
-                        .message()
-                        .contentType(MediaType.APPLICATION_JSON_VALUE)
-                        .body("{\n" +
-                                "\"color\": \"" + color + "\",\n" +
-                                "\"height\": " + height + ",\n" +
-                                "\"material\": \"" + material + "\",\n" +
-                                "\"sound\": \"" + sound + "\",\n" +
-                                "\"wingsState\": \"" + wingsState + "\"\n" + "}"));
-    }
-
-    public void deleteDuck(TestCaseRunner runner, String id) {
-        runner.$(
-                http()
-                        .client("http://localhost:2222")
-                        .send()
-                        .delete("/api/duck/delete")
-                        .queryParam("id", id));
-    }
-
-    public void getDuckId(TestCaseRunner runner) {
-        runner.$(
-                http()
-                        .client("http://localhost:2222")
-                        .receive()
-                        .response(HttpStatus.OK)
-                        .message()
-                        .contentType(MediaType.APPLICATION_JSON_VALUE)
-                        .extract(fromBody().expression("$.id", "duckId"))
-        );
-    }
-
-    public void validateResponse(TestCaseRunner runner, String responseMessage) {
-        runner.$(
-                http()
-                        .client("http://localhost:2222")
-                        .receive()
-                        .response(HttpStatus.OK)
-                        .message()
-                        .contentType(MediaType.APPLICATION_JSON_VALUE)
-                        .body(responseMessage));
-    }
-}
