+++ src/test/java/autotests/clients/DuckActionsClient.java
@@ -1,0 +1,148 @@
+package autotests.clients;
+
+import autotests.EndpointConfig;
+import com.consol.citrus.TestCaseRunner;
+import com.consol.citrus.http.client.HttpClient;
+import com.consol.citrus.message.MessageType;
+import com.consol.citrus.testng.spring.TestNGCitrusSpringSupport;
+import com.consol.citrus.validation.json.JsonPathMessageValidationContext;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.test.context.ContextConfiguration;
+
+import static com.consol.citrus.dsl.MessageSupport.MessageBodySupport.fromBody;
+import static com.consol.citrus.http.actions.HttpActionBuilder.http;
+
+@ContextConfiguration(classes = EndpointConfig.class)
+public class DuckActionsClient extends TestNGCitrusSpringSupport {
+
+    @Autowired
+    protected HttpClient duckService;
+
+    public void createDuck(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
+        runner.$(
+                http()
+                        .client(duckService)
+                        .send()
+                        .post("/api/duck/create")
+                        .message()
+                        .contentType(MediaType.APPLICATION_JSON_VALUE)
+                        .body("{\n" +
+                                "\"color\": \"" + color + "\",\n" +
+                                "\"height\": " + height + ",\n" +
+                                "\"material\": \"" + material + "\",\n" +
+                                "\"sound\": \"" + sound + "\",\n" +
+                                "\"wingsState\": \"" + wingsState + "\"\n" + "}"));
+    }
+
+    public void deleteDuck(TestCaseRunner runner, String id) {
+        runner.$(
+                http()
+                        .client(duckService)
+                        .send()
+                        .delete("/api/duck/delete")
+                        .queryParam("id", id));
+    }
+
+    public void getDuckId(TestCaseRunner runner) {
+        runner.$(
+                http()
+                        .client(duckService)
+                        .receive()
+                        .response(HttpStatus.OK)
+                        .message()
+                        .contentType(MediaType.APPLICATION_JSON_VALUE)
+                        .extract(fromBody().expression("$.id", "duckId"))
+        );
+    }
+
+    public void validateResponse(TestCaseRunner runner, String responseMessage) {
+        runner.$(
+                http()
+                        .client(duckService)
+                        .receive()
+                        .response(HttpStatus.OK)
+                        .message()
+                        .contentType(MediaType.APPLICATION_JSON_VALUE)
+                        .body(responseMessage));
+    }
+
+    public void validateErrorResponse(TestCaseRunner runner, HttpStatus expectedStatus, String expectedMessage) {
+        runner.$(
+                http()
+                        .client(duckService)
+                        .receive()
+                        .response(expectedStatus)
+                        .message()
+                        .contentType(MediaType.APPLICATION_JSON_VALUE)
+                        .body("{\n" + "  \"message\": \"" + expectedMessage + "\"\n" + "}"));
+    }
+
+    public void validateResponseJsonPath(TestCaseRunner runner, JsonPathMessageValidationContext.Builder body) {
+        runner.$(
+                http()
+                        .client(duckService)
+                        .receive()
+                        .response(HttpStatus.OK)
+                        .message()
+                        .type(MessageType.JSON)
+                        .validate(body)
+        );
+    }
+
+    public void duckSwim(TestCaseRunner runner, String id) {
+        runner.$(
+                http()
+                        .client(duckService)
+                        .send()
+                        .get("/api/duck/action/swim")
+                        .queryParam("id", id));
+    }
+
+    public void quack(TestCaseRunner runner, String id, String repetitions, String quacks) {
+        runner.$(
+                http()
+                        .client(duckService)
+                        .send()
+                        .get("/api/duck/action/quack")
+                        .message()
+                        .queryParam("id", id)
+                        .queryParam("repetitionCount", repetitions)
+                        .queryParam("soundCount", quacks));
+    }
+
+    public void propertiesDuck(TestCaseRunner runner, String id) {
+        runner.$(
+                http()
+                        .client(duckService)
+                        .send()
+                        .get("/api/duck/action/properties")
+                        .queryParam("id", id));
+    }
+
+    public void duckFly(TestCaseRunner runner, String id) {
+        runner.$(
+                http()
+                        .client(duckService)
+                        .send()
+                        .get("/api/duck/action/fly")
+                        .queryParam("id", id));
+    }
+
+    public void updateDuck(TestCaseRunner runner, String id, String color, double height, String material, String sound, String wingsState) {
+        runner.$(
+                http()
+                        .client(duckService)
+                        .send()
+                        .put("/api/duck/update")
+                        .message()
+                        .queryParam("id", id)
+                        .queryParam("color", color)
+                        .queryParam("height", String.valueOf(height))
+                        .queryParam("material", material)
+                        .queryParam("sound", sound)
+                        .queryParam("wingsState", wingsState)
+        );
+    }
+}
