+++ src/test/java/autotests/tests/duckController/CreateTest.java
@@ -1,0 +1,26 @@
+package autotests.tests.duckController;
+
+import autotests.clients.DuckActionsClient;
+import com.consol.citrus.TestCaseRunner;
+import com.consol.citrus.annotations.CitrusResource;
+import com.consol.citrus.annotations.CitrusTest;
+import org.testng.annotations.Optional;
+import org.testng.annotations.Test;
+
+import static com.consol.citrus.validation.json.JsonPathMessageValidationContext.Builder.jsonPath;
+
+public class CreateTest extends DuckActionsClient {
+    @Test(description = "Проверка того, что уточка создалась с материалом rubber")
+    @CitrusTest
+    public void createRubberDuck(@Optional @CitrusResource TestCaseRunner runner) {
+        createDuck(runner, "yellow", 0.03, "rubber", "quack", "ACTIVE");
+        validateResponseJsonPath(runner, jsonPath().expression("$.material", "rubber"));
+    }
+
+    @Test(description = "Проверка того, что уточка создалась с материалом wood")
+    @CitrusTest
+    public void createWoodDuck(@Optional @CitrusResource TestCaseRunner runner) {
+        createDuck(runner, "yellow", 0.03, "wood", "quack", "ACTIVE");
+        validateResponseJsonPath(runner, jsonPath().expression("$.material", "wood"));
+    }
+}
