+++ src/test/java/autotests/tests/duckController/DeleteTest.java
@@ -1,0 +1,19 @@
+package autotests.tests.duckController;
+
+import autotests.clients.DuckActionsClient;
+import com.consol.citrus.TestCaseRunner;
+import com.consol.citrus.annotations.CitrusResource;
+import com.consol.citrus.annotations.CitrusTest;
+import org.testng.annotations.Optional;
+import org.testng.annotations.Test;
+
+public class DeleteTest extends DuckActionsClient {
+    @Test(description = "Проверка того, что уточка удалилась")
+    @CitrusTest
+    public void successfulDelete(@Optional @CitrusResource TestCaseRunner runner) {
+        createDuck(runner, "yellow", 0.03, "rubber", "quack", "ACTIVE");
+        getDuckId(runner);
+        deleteDuck(runner, "${duckId}");
+        validateResponse(runner, "{\n  \"message\": \"Duck is deleted\"\n}");
+    }
+}
